Add baseline for facade pattern
